pyproject.toml
[project]
name = "duck_quant"
version = "0.1.0"
description = "Data pipeline with Google Sheets, DuckDB, etc."
requires-python = ">=3.10,<3.12"

# Dependencies
dependencies = ["pandas"]

# Dependency groups
[project.optional-dependencies]
extract = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "requests",
]
transform = [
    "duckdb",
    "dbt-core",
    "dbt-duckdb",
]
orchestration = [
    "apache-airflow",
    "astronomer-cosmos",
    "flask==2.1.3",
    "werkzeug==2.3.7"
]
dev = [
    "pytest",
    "black",
    "pre-commit"
]

[tool.setuptools]
packages = ["orchestration"]
